In browse page move rescore tool into output list
