+++ app/components/OutputReport.tsx
@@ -1,6 +1,7 @@
 import { useMemo } from "react";
 import type { DirectoryItem } from "~/bartender-client";
 import { ListFiles } from "./ListFiles";
+import { getLastCaprievalModule } from "~/tools/shared";
 
 export function files2modules(files: DirectoryItem) {
   if (!files.children) {
@@ -9,6 +10,12 @@ export function files2modules(files: DirectoryItem) {
   const analyisRoot = files.children.find((i) => i.name === "analysis");
 
   const nonmodules = new Set(["analysis", "data", "log", "traceback"]);
+  let lastCaprievalModuleIndex = -1;
+  try {
+    lastCaprievalModuleIndex = getLastCaprievalModule(files);
+  } catch (error) {
+    // ignore
+  }
   return files.children
     .filter((i) => !nonmodules.has(i.name))
     .filter((i) => !i.name.endsWith("_interactive"))
@@ -22,6 +29,7 @@ export function files2modules(files: DirectoryItem) {
         name,
         output,
         report,
+        isLastCaprieval: parseInt(id) === lastCaprievalModuleIndex,
       };
     });
 }
@@ -63,6 +71,11 @@ export const OutputReport = ({
                     rel="noreferrer"
                     href={`/jobs/${jobid}/tools/reclustrmsd/${module.id}`}
                   >
+                    🔧
+                  </a>
+                )}
+                {module.isLastCaprieval && (
+                  <a title="Rescore" href={`/jobs/${jobid}/tools/rescore`}>
                     🔧
                   </a>
                 )}
